fix demo shell
