+++ ng2-components/ng2-activiti-tasklist/src/components/task-filters.component.css
